Cleaned up code from main and added seperate module/files for the sub systems
